Else conditions

+++ remotecollaboration.py
@@ -81,9 +81,9 @@ def loop():
     renderDropdown('Local Branches', current_local_branch, local_branch_names)
     renderDropdown('Remote Branches', current_remote_branch, remote_branch_names)
     if imgui_python.ImGui_Button(ctx, 'Delete Selected Local Branch'):
-      deleteSelectedBranch()
+      deleteSelectedBranch(False)
     if imgui_python.ImGui_Button(ctx, 'Delete Selected Remote Branch'):
-      deleteSelectedBranch()
+      deleteSelectedBranch(True)
 
     (show_textinput, message) = imgui_python.ImGui_InputText(ctx, 'Commit message', commit_message[0])
     if show_textinput:
@@ -146,22 +146,22 @@ def checkout(branch: str):
     repo.heads[ref.remote_head].checkout()
     current_local_branch[0] = str.split(branch,'/')[1]
 
-def deleteSelectedBranch():
+def deleteSelectedBranch(isRemote: bool):
   # Checkout default branch to avoid errors
   repo.heads.main.checkout()
-  try:
+  if isRemote:
     # Store menu binding for deletion
     deleting_head = current_local_branch[0]
     repo.delete_head(deleting_head)
     local_branch_names.remove(deleting_head)
-  except:
+  else:
     # Store menu binding for deletion
     deleting_branch = str.split(current_remote_branch[0],'/')[1]
     # Set menu binding to default branch
     current_remote_branch[0] = origin.refs.main.name
     with repo.git.custom_environment(GIT_SSH_COMMAND=git_ssh_cmd):
-        origin.push(refspec=(":%s" % deleting_branch))
-        updateBranchList()
+      origin.push(refspec=(":%s" % deleting_branch))
+      updateBranchList()
   # Set menu binding to default branch
   current_local_branch[0] = repo.active_branch
   # Open the project for default branch
